refactor(orchestrator): split tests into backend-agnostic + docker-specific (#179)
The pre-phase-3-8 test file tested the backend-agnostic state machine
through Docker-specific fixtures, meaning no coverage in the
process_backend build variant. Split into an untagged orchestrator_test.go
(42 tests using a lightweight fakeServerFactory) and a trimmed
clone_docker_test.go (13 tests calling factory methods directly). Removes
all test-only shim methods on *Orchestrator.
